wip: add some aft functions
